+++ tests/testthat/test-negate.R
@@ -1,0 +1,5 @@
+test_that("negate() works", {
+  different <- negate(identical)
+  expect_identical(formals(different), formals(identical))
+  expect_true(different(different, identical))
+})